+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+- fixed bug where if first arguments weren't flags they were skipped
+
 ## v1.0.0
 - parses booleans, strings, arrays
 - collects extra arguments as `_`